% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_n_days_rep}
\alias{last_n_days_rep}
\title{Returns the last week absolute numbers}
\usage{
last_n_days_rep(data, n = 7)
}
\arguments{
\item{data}{\code{data.frame} of covid19 data}

\item{n}{\code{integer}, number of days to consider, default 7}
}
\value{
\code{data.frame}
}
\description{
Returns the last week absolute numbers
}
\examples{
df <- read_data("case_time_series.csv","covid_data")
dat <- process_data(df)
last_n_days_rep(dat)
}
\seealso{
\code{\link[=formatC]{formatC()}}
}
